Stratabase fixes: Primarily list representation (#44)
Major improvements to Stratabase in terms of list representation. Not only was it broken before, but logically the insertion style did not make sense. The real goal is to have observable element adds/removes representation in the list access which has now been achieved. The new style has lists directly stored in each layer, where the topmost layer represents the entire list, not some element override structure.

This is easier to understand, follows the Stratabase model better, and most importantly, works.